Users and messages - initial work
Implementing:

- State management for users lookup cache. This is a dynamically
populated cache when a message is from a user that the cache doesn't
already know about. The MessageList component causes the cache to be
updated, which the Message component can then use for display.

- State management for active user. This is the very beginnings of an
active user state object. It is currently just used for identifying if a
message is from "You" or not and for populating the ActiveUser
component.

- Also added the beginnings on layout for message display