Message collections
This commit introduced the concept of message collections, which is
primarily used for displaying messages in logical groups.

Some refactoring was required due to the fact that message collections
come into play between the MessageList and Message components, MessageCollections are generated from the MessageList and then Message components are used to display the actual messages.

There is also work included to display messages in collections as per
the mockups

This commit also includes tests for the collection generation logic