Message display
Adding message display logic and formatting, and associated unit tests.

The message display is dictated by the user sending the collection
(which comprises one or more messages) and the initiator of the query.
So here I have added the logic to calculate this and the appropriate
styling to make it happen.