Query selection, UI changes and message population
This commit adds routing to determine the active query. When a user
clicks on a query, the route is changed to /query/<query_id>, this route
is then used by the QueryHeader and MessageList components. The
QueryHeader component derives the title for display from the queryList
state property. The MessageList component populates itself from the API.

Styling has also been added to provide the user with feedback as to
which query is active.