Retrieve and display asociated user information
This commit adds the ability for components to request users and then
display it, including user avatars.

We also keep track of ongoing requests for users. The
idea being that if multiple components are requesting user details from
the API, users that have already been requested, or are in the process
of being requested will not be requested again

- Added react-avatar package
- Added actions and reducers for fetching and storing users
- Added the display of users and their avatars to the appropriate
components
- Added unit tests for all above functionality