Keep query list in sync with active query
When a message is added, edited or deleted in the active query, we need
the query summary in the querylist to reflect those changes. This commit
adds this. Currently, we dispatch an action that causes the query
summary to refresh itself from the API, this is a tiny request which
will only happen infrequently. Nevertheless, we may want to move this to
be part of the intended websocket message that includes unread count.

Includes:
- Action and reducers for updating a query
- Modifications to dispatch actions
- Unit test modifications