Query searching and new query
This commits adds the ability for the user to search for queries by
title. Matching queries are displayed along with (if appropriate) the
currently active query.

If no queries are returned from a search, a button is displayed allowing
the creation of a new query. Also, the search bar now contains a
persistent button that allows the display of a new query.

Includes:

- Actions & reducers for setting search string
- UI additions and modifications to accept and display search results
- Unit tests