Calculated folders
The "All queries" and "Unread" folders are "virtual" folders in that
their contents are derived from the message status of all the queries.

Currently we have no message status, so these folders do not have any
contents, and their message counts are just placeholders until we do.
The "All queries" message count currently reflects the total number of
queries.

Includes:

- CalculatedFolders component including minimal behaviour
- Minimal unit tests