Fix folder_id => folder renaming
The property of a query object referring to its folder has been renamed
from 'query_id' to 'query', to reflect the fact that the foreign key in
the database is no longer an ID, it joins with the folder's code.

This commit makes the required changes in our unit tests.