Fix message update failure handling
Prior to sending an API request for an object update, we pre-emptively
update our local state to reflect the change, thereby giving a more
responsive experience.

If a message update failed we were incorrectly just removing the
offending message from the UI.

Working on the query update highlighted that we should be reverting to
the original, unmodified, object if the API request fails. This
behaviour is what happens with query updates.

This commit adds this behaviour to messages also.