Switch from homegrown Popover to library
The homegrown <Popover> was showing limitations, mostly due to stacking
context, z-index and overflow. In Chrome alone it was apparent that
there were edge cases that were going to catch us out. These issues
could be resolved, but it became obvious that this was going to create
more technical debt than using a mature Popover component library.

Consequently I have adopted react-tiny-popover. This commit removes our
own Popover and makes changes to the ColourPicker that used it.