Query / label assignment
This commit adds the ability for one or more labels to be attached and
detached from queries.

Includes:
- Generic OptionPicker component which is passed an array of options and
an array of selected options. It is then possible for the user to select
and de-select options, which is then reflected in the selected options
array
- LabelPicker component which uses the OptionPicker component to display
a Popover of all the available query labels, indicating which are
already attached to the given query. The user can then select &
de-select from the list
- Redux action for toggling label / query relationships
- Redux reducer for toggling label / query relationships
- New unit tests for new components
- Modified existing unit tests as required