Follow up on label management
There was a bug whereby if a label was deleted when it was attached to
one or more queries, the app would crash.

The fix implemented here is for the queries reducer to catch the
DELETE_LABEL_SUCCESS action and remove that label from all queries that
use it. It is sufficient to do that, rather than refetch all the
relevant queries because this is a very small atomic change that we can
be confident we are replicating accurately in the frontend state.

Added unit test for the new queries reducer