Query bulk operations
This commit adds the ability for queries to be deleted or labelled in
bulk.

We have switched to the approach of *always* performing these actions in
bulk, even if we're only doing it on a single query. This results in not
having to maintain two code paths for what are essentially the same
operation.

Includes:

- Modified relevant components to work with multiple queries, rather
than a single query. Including new wrapping components where required
- Modified QueryBulk component to provide the ability to:
  - Select / de-select all queries
  - Move selected queries to "BIN" folder
  - Label selected queries
- New useFilters custom hook which provides the current filtered state
of the queries
- Modified and created unit tests as required