Authentication and authorisation
This commit adds authentication and authorisation support to the
frontend of the EMS. This allows the user to choose the method they wish
to use to authenticate, from the list of available methods.
Authenticate, then use the application whilst remaining authenticated,
the log out when they are done.

It also includes a few minor, unrelated changes, detailed below.

Includes:

- (Unrelated) Move FontAwesome icon initialisation to its own module
- We now block mounting of the app until we have a valid JWT. The couple
of hundred milliseconds this take is unnoticible and ensures we have a
JWT for any subsequent API calls
- Added Login component that lists available auth types and provides
links to them
- Created useActiveUser custom hook - provides data and functions
relating to the active user
- Created useMountable custom hook - provides the consuming component
the ability to know whether it can mount, depending on the availability
of a JWT
- Created Redux store for activeUser
- CORS mode on XHR requests is now only enabled when in development.
This corresponds with a similar change in the API
- Switched from using the Fetch API to the Axios library. The main
impetus for this is twofold. Fetch handles requests that fail (say, with
a 401) as successes and it's necessary to manually check the response
code to know whether you have to intervene, which is tedious. Also, we
needed a request / response interceptor mechanism in order to handle JWT
grabbing from a response, cache, and attach to a request, Axios provides this.
- All API requests are now performed via an EmsApi class singleton
instance. This class handles the token interaction as mentioned above,
it also manages automatically getting a new JWT when the current one is
about to expire.
- Modified unit tests in accordance with above changed