Message collections should have a stable ID
We calculate message collections on the fly. However, I was using a
random value for the collection key. This was a terrible idea. Because
each collection's key changed every time a change occurred, the
collections we recalculated but, because all their keys had changed,
React re-rendered the entire message list. Doh!

We now use a combination of data that will be unique to the collection
to form its key.