Add queries.activeQuery state
There are a few places where we need to know about the active query,
thus far we've been using router params. But the upcoming websocket
functionality needs to know it but doesn't have access to the router, so
we're moving it into state, which is where it should be anyway.

Includes:

- New action & reducer for queries.activeQuery
- Update state when route param changes in ActiveQuery.js