Store user role in activeUser state
The API call to /activeUser returns a complete user object. Thus far
we've only been storing the user ID in the activeUser state. We now
store the role, this will enable us to easily establish a user's role to
determine what UI to display.

It's possible for a nefarious actor to modify the state to try to elevate
their role, this will enable them to see bits of the elevated UI, but it
won't be populated and they won't be able to carry out any actions since
their user won't be privileged to access the necessary API routes