Add API request in progress cache
We need to monitor requests that are in progress in order to make sure
we don't send multiple requests for the same thing. This is most
apparent with components that use the useActiveUser hook. Since hooks
are not singletons, whenever a component uses this hook it tries to
populate the state, if the state isn't already populated. However, when
the app first mounts, multiple components try to do this concurrently,
only one should. So, now, only the first attempt at the request actually
happens.

This is implemented in the EmsApi class, so we benefit from this
behaviour for all requests.

Includes:

- Added inProgress array that is updated when new requests are made and
are completed, this is checked when a new request is about to be made to
make sure it is not made again