Refactor message creation state to use Redux
Previously we had our message creation state (the ID and text of the
message being created / edited) held in the ActiveQuery component, it was
then passed down to to QueryEntry component and, more problematically,
down via MessageList > MessageCollection to the Message component. This
had performance impacts on QueryEntry due to the common parent
re-rendering, causing the re-render of all descendants.

This was causing very slow performance typing a message when the
message list was large (tested with ~300 messages in a query).

This commit refactors this state out to Redux. This fixes the
performance problems, and also prepares us for the next thing which is a
self-contained message entry component when creating a new query.

Includes:

- New activeMessage Redux state
- Types, Actions and Reducers for manipulating the activeMessage state
- Adaptation of components that previously used the state held in
ActiveQuery to use the new Redux state