Do not allow app mounting when with no auth
There was a race condition whereby the app could start mounting and
rendering before the first API call was made. We did not know at that
point whether the user was authorised so it would result in the app
starting to appear then the user being redirected to the /login route.
There was no security implications, it just looked bad. This commit
fixes this.

Includes:
- We now have a new 'hasAuth' property returned from useMountable, this
is used to determine if the app has retrieved an authorisation token
from the API. We use this property to block rendering of the app until
it is true.