Fix race condition on logout
There was a race condition when the user logged out. Part of the logout
process was to reset the application state and redirect user to the
/login route. The problem was that the redirect was happening before the
state change had finished, resulting in an error. The easiest fix for
this is to move the resetting of the state to the mounting of the Login
component, which is arguably where it should have been in the first
place.

Includes:
- Move logoutUser() dispatch to Login component
- Add missing translation for "Log out" text