Fix 'mounted' bug and re-enable scrolling
This commit fixes a bug in the hook that maintains, amongst other
things, the array of mounted messages. We were adding both a new
message's temporary ID and then it's final ID, this was resulting in the
same message appearing twice in the mounted array. The code that used
the mounted length matching the message length as an indicator of
whether it should do something was consequently broken.

It also re-enables message scrolling upon query mount. We use the
mounted list to establish when we should scroll to the bottom, hence the
above fix.