Move away from components using 'match'
Up until now, if a component wanted to know what query it was handling,
it would check the 'match' prop supplied by withRouter. This wasn't
ideal because we want to be able to manipulate the active query.

So now, we have a 'useActiveQuery' hook which initially populates its
state using the 'match' param returned from the useParams hook provided
by react-router-dom. Whenever a new query loads and the param changes it
updates, but we can also manipulate it.

This will come in useful when we want to programmatically remove the
active query.

Includes:
- Add new useActiveQuery hook
- Modify existing components that used 'match' to use the new hook
instead