Move styling of active query to Query component
Until now the "active" styling of the query was done via
react-router-dom adding an 'active' class to the query <li> element.
This is no longer appropriate as we're decoupling our "active" query and
the :queryId URL param (i.e. our active query may be different to the ID
in the URL, specifically we might want to set the active query to null
when the user is searching or filtering).

Includes:
- Removes active style from index.scss
- Adds "am I active" check and appropriate style to the Query module
- Switch from NavLink to Link component