Modify query action button layout
The query list container did not scroll when the number of queries
exceeded its height. Adding overflow-y: scroll introduced a world of
pain. Until now, we had been relying on the query action button being
able to overflow on the x axis over the acitve query container. This
allowed the button to extend outside of its container when hovered over.

Adding the overflow to the query list completely breaks this. I've spent
a _very long time_ trying to find out how to fix it. I think the problem
may be something to do with stacking contexts, but I have been unable to
find a fix. I'm now beginning to wonder if it's even possible.

Therefore, this commit side-steps the problem by making the query action
button extend to the left, instead of the right. This is not ideal, but
it works and doesn't negatively impact the layout of the query.

I've posted a question on SO in the hope that someone will be able to
see a solution that I can't! At which point we will be able to revert to
the previous layout. Fingers crossed!