Don't lose destination following authentication
If an unauthenticated user is attempting to go to a deep link, for
example, /query/123 - the location they were attempting to go to, prior
to being redirected to login, was not being stored. Consequently,
following authentication, they would just be taken to the homepage.

This commits adds the storage of the destination. Includes:

- New Destination class, which manages the storage, fetching and
clearing of a 'dest' cookie that holds an intended destination. I opted
for a class rather than a React hook because I needed to use it in the
EmsApi class.
- When a 401 response is intercepted, we store the destination
- The router then looks for a destination cookie and, if one exists,
uses it